gps packet data type optimization

+++ End Device/packet-test/packet.h
@@ -5,8 +5,8 @@
 class Packet {
 public:
     byte direction = 0;
-    double latitude = 0;
-    double longitude = 0;
+    int32_t latitude = 0;
+    int32_t longitude = 0;
     int16_t temperature = -99;
     int16_t humidity = -99;
     int8_t battery = 0;
